Regroup the database layer under Core/Database/Access
Move record access (the Database protocol, readers, writers, queries, locator, and sender) into a Core/Database/Access subfolder, add RecordValueConvertible, and drop the standalone RecordID, RecordChunk, RecordCursor, RecordErrors, and RecordProtocols files. Mirror the move in the matching test folder.
