+++ Sources/Scout/Core/Database/Access/Database.swift
@@ -1,0 +1,13 @@
+//
+// Copyright 2026 Mikhail Kasianov
+//
+// Use of this source code is governed by an MIT-style
+// license that can be found in the LICENSE file or at
+// https://opensource.org/licenses/MIT.
+
+import Foundation
+
+typealias Database = DatabaseReader & DatabaseWriter
+
+typealias DatabaseReader = RecordLocator & MetricSeriesReader & ActivityReader & Sendable
+typealias DatabaseWriter = RecordWriter & Sendable
