+++ Sources/Scout/Core/Database/Access/RecordWriter.swift
@@ -12,10 +12,7 @@ protocol RecordWriter: Sendable {
     func write(records: [Record]) async throws
 }
 
-extension Array {
-    func chunked(into size: Int) -> [[Element]] {
-        stride(from: 0, to: count, by: size).map {
-            Array(self[$0..<Swift.min($0 + size, count)])
-        }
-    }
+struct RecordConflictError: LocalizedError {
+    let serverRecord: Record
+    let errorDescription: String? = "The record was changed on the server"
 }
